pyproject.toml
[tool.poetry]
name = "ecommerce"
version = "1.0.0"
description = ""
authors = ["test <test@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^4.2.0"
djangorestframework-simplejwt = "^5.3.1"
django-environ = "^0.11.2"
pillow = "^10.2.0"
django-redis = "^5.4.0"
pytest = "^8.0.1"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
